+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.1...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.2...HEAD) - TBD
+
+## [4.21.2](https://github.com/schemathesis/schemathesis/compare/v4.21.1...v4.21.2) - 2026-06-09
 
 ### :bug: Fixed